+++ src/launcher/features/apps/App/CreateShortcut.tsx
@@ -8,17 +8,14 @@ import React from 'react';
 import Dropdown from 'react-bootstrap/Dropdown';
 
 import { createDesktopShortcut } from '../../../../ipc/createDesktopShortcut';
-import { useLauncherSelector } from '../../../util/hooks';
-import { DisplayedApp, getIsAnAppInProgress } from '../appsSlice';
+import { DisplayedApp, isInProgress } from '../appsSlice';
 
 const CreateShortcut: React.FC<{ app: DisplayedApp }> = ({ app }) => {
-    const isAnAppInProgress = useLauncherSelector(getIsAnAppInProgress);
-
     if (!app.isInstalled) return null;
 
     return (
         <Dropdown.Item
-            disabled={isAnAppInProgress}
+            disabled={isInProgress(app)}
             title="Create a desktop shortcut for this app"
             onClick={() => createDesktopShortcut(app)}
         >