+++ src/launcher/features/apps/App/InstallApp.tsx
@@ -7,21 +7,19 @@
 import React from 'react';
 import Button from 'react-bootstrap/Button';
 
-import { useLauncherDispatch, useLauncherSelector } from '../../../util/hooks';
+import { useLauncherDispatch } from '../../../util/hooks';
 import { installDownloadableApp } from '../appsEffects';
-import { DisplayedApp, getIsAnAppInProgress } from '../appsSlice';
+import { DisplayedApp, isInProgress } from '../appsSlice';
 
 const InstallApp: React.FC<{ app: DisplayedApp }> = ({ app }) => {
     const dispatch = useLauncherDispatch();
-    const isAnAppInProgress = useLauncherSelector(getIsAnAppInProgress);
-
     if (app.isInstalled) return null;
 
     return (
         <Button
             variant="outline-secondary"
             title={`Install ${app.displayName}`}
-            disabled={isAnAppInProgress}
+            disabled={isInProgress(app)}
             onClick={() => dispatch(installDownloadableApp(app))}
         >
             {app.progress.isInstalling ? 'Installing...' : 'Install'}