+++ Changelog.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+### Changed
+
+-   Installing or upgrading an app does not prevent working with other apps in
+    the list of apps any longer.
+
 ### Removed
 
 -   Removed pc-ble-driver as the Bluetooth Low Energy app now contains it